Fix: refactoring + linting errors
