+++ brain_games/cli.py
@@ -4,7 +4,11 @@ import prompt
 
 
 def welcome_user():
-    """User acquaintance."""
+    """User acquaintance.
+
+    Returns:
+        user's name.
+    """
     name = prompt.string('May I have your name? ')
     print('Hello, {0}!'.format(name))
     return name
