+++ setup.cfg
@@ -14,9 +14,9 @@ isort-show-traceback = True
 ignore =
 
 per-file-ignores =
-  # allowing prints (421)
+  # allowing prints (421) and standard pseudo-random generators (S311)
   hexlet_python_package/scripts/*.py: WPS421
-  brain_games/*.py: WPS421
+  brain_games/*.py: WPS421, S311
 
 [tool:pytest]
 norecursedirs = __pycache__
